Padding, numProcessors and problem row indices
Added algorithm to calculate how much to pad the problem array with
"dummy" rows, and to select optimum number of processors to cover all
the required rows.

Also changed how the problem row indices are added to the problem array.
This is now not included in the problem.c file, and is instead done in
main. This separates concerns and hopefully improves readability.