Fixed solve at different times issues
This was due to the fact that a thread would:
 - Relax rows on values
 - Gather into newValues
 - Update values with newValues, and compare to see if anything changed

The problem with this is that every thread then had a different 'values'
array, meaning some threads were being markes as solved and others were
not. This meant that these threads looped again and hung on the gather.
This was fixed incorrectly by splitting the comm every time, but this
fix was wrong and was resulting in incorrect answers. The actual fix was
to have the relaxation stage work on newValues, so that the original
values array is stored and is the same for all processors.