Fixed padding rows issue
We need to give solve both the problem size (rows) and the total rows,
so that relaxRows and updateRows can only loop over the problem, and
ignore padding rows.

This bug occurred when there was more than one, say 2, padding rows.
The third last row in the array was considered not an edge values
(even though it actually was) and hence was averaged using the 0 values
of the padding rows, producing incorrect results.

This should also slightly improve efficiency as we only loop over values
we need to loop over.