Removed row indices in problem array method
This is not needed. Allgatherv places the data received from process i
at displs[i]. The result of this is that all data will be
received/ordered in the correct order, meaning these indices are no
longer needed as they were only there to maintain order. This is now
more readable, cleaner and should also be faster.

Also renamed values -> newValues for readability as this is what is
passed in to this function.